modules/knowledge-measure-businness/requirements.txt
agent-base==0.23.0
agents-api~=0.27.0
aio-pika==7.1.2
aiohttp==3.7.4.post0
aiosignal==1.2.0
asynctest==0.13.0
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
company-name==0.18.1
coverage
data-source-api==0.19.0.74
date-agent-api~=0.27.0
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
flake8-junit-report
Flask==2.0.2
frozenlist==1.2.0
geo-agent-api~=0.27.0
googleapis-common-protos==1.53.0
grpcio~=1.46.3
grpcio-health-checking~=1.46.3
grpcio-reflection~=1.46.3
grpcio-status~=1.46.3
grpcio-tools~=1.46.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lz4==3.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
notebook==6.4.6
numpy==1.21.5
organizationname-agent-api~=0.27.0
organization-name-knowledge==0.9.0
packaging==21.3
pathspec==0.9.0
phonetics==1.0.5
platformdirs==2.4.1
pluggy==1.0.0
protobuf~=3.19.4
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov
python-consul==1.1.0
python-consul2==0.1.5
PyYAML==5.4.1
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
tox==3.24.5
tox-docker==4.0.0a1
typed-ast==1.5.1
typing_extensions==4.0.1
Unidecode==1.2.0
urllib3==1.26.7
virtualenv==20.13.0
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
